Update expense and design fixes

+++ src/app/feed/feed.component.ts
@@ -15,14 +15,53 @@ export class FeedComponent implements OnInit {
 
   expenses = new Observable<Expense[]>();
 
+  editEntity = {
+    id: 0,
+    show: false
+  }
+
   ngOnInit(): void {
     this.expenses = this.expenseService.expensesList;
   }
 
-  remove(idExpense: string, ind: number){
+  remove(idExpense: string, ind: number, title: string){
+
+    if (confirm("Are you sure to delete "+title+" expense?")) {
+      this.expenseService.deleteExpense(idExpense, ind);
+    }
     
-    this.expenseService.deleteExpense(idExpense, ind);
+  }
+
+  update(idExpense: string, ind: number, title: string, category: string, price: string){
+    console.log(`Editing expense with id: ${idExpense} [${ind} in array] with ${title} ${category} ${price}`);
+
+    const tmpObj: Expense = {
+      id: idExpense,
+      category,
+      title,
+      price: Number.parseFloat(price)
+    }
     
+    this.expenseService.updateExpense(tmpObj, ind);
+
+    this.toggleEdit(-1);
+
+  }
+
+  toggleEdit(ind: number){
+
+    if(this.editEntity.show){
+      if(this.editEntity.id == ind){
+        this.editEntity.show = false;
+      }else{
+        this.editEntity.id = ind;
+      }
+    }else{
+      this.editEntity.id = ind;
+      this.editEntity.show = true;
+    }
+
+
   }
